requirements.txt
# Django Core
Django==4.2.15
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
mysqlclient==2.2.0
pymysql==1.1.0

# Redis Cache
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# API Documentation
drf-spectacular==0.27.0

# Environment
python-decouple==3.8

# Excel/Reports
openpyxl==3.1.2
pandas==2.1.4
xlsxwriter==3.1.9

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
requests==2.31.0

# System Monitoring (NEW - for system health endpoints)
psutil==7.1.0

# Task Queue (optional)
# celery==5.3.4
# redis==5.0.1
